# openapi/ResponseExamples.Api.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "Product API",
    "description": "A RESTful API for managing products with full CRUD operations",
    "version": "v1"
  },
  "paths": {
...
    "/api/errors/validation-error": {
      "post": {
        "tags": [
          "Error Examples"
        ],
        "summary": "Demonstrates a 400 Bad Request with validation errors",
        "description": "Returns a ValidationProblemDetails response showing multiple validation failures. Pass invalid data to trigger validation errors.",
        "operationId": "ValidationError",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "oneOf": [
                  {
                    "nullable": true
                  },
                  {
                    "$ref": "#/components/schemas/ValidationRequest"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/errors/not-found/{id}": {
      "get": {
        "tags": [
          "Error Examples"
        ],
        "summary": "Demonstrates a 404 Not Found error",
        "description": "Returns a ProblemDetails response when a resource is not found",
        "operationId": "NotFoundError",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Any ID value will trigger a 404 response",
            "required": true,
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/errors/conflict": {
      "post": {
        "tags": [
          "Error Examples"
        ],
        "summary": "Demonstrates a 409 Conflict error",
        "description": "Returns a ProblemDetails response when there's a conflict with the current state, such as attempting to create a duplicate resource",
        "operationId": "ConflictError",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "oneOf": [
                  {
                    "nullable": true
                  },
                  {
                    "$ref": "#/components/schemas/ConflictRequest"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/errors/unprocessable": {
      "post": {
        "tags": [
          "Error Examples"
        ],
        "summary": "Demonstrates a 422 Unprocessable Entity error",
        "description": "Returns a ProblemDetails response when the request is well-formed but semantically incorrect",
        "operationId": "UnprocessableEntityError",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "oneOf": [
                  {
                    "nullable": true
                  },
                  {
                    "$ref": "#/components/schemas/UnprocessableRequest"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/errors/server-error": {
      "get": {
        "tags": [
          "Error Examples"
        ],
        "summary": "Demonstrates a 500 Internal Server Error from an exception",
        "description": "Throws an exception that gets caught by the ProblemDetails middleware and converted to a standardized error response",
        "operationId": "ServerError",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ConflictRequest": {
        "type": "object",
        "properties": {
          "identifier": {
            "type": "string",
            "nullable": true
          }
        },
        "description": "Request model for conflict error demonstration"
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
      "UnprocessableRequest": {
        "type": "object",
        "properties": {
          "startDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "endDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        },
        "description": "Request model for unprocessable entity demonstration"
      },
...
      "ValidationProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        }
      },
      "ValidationRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "age": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "format": "int32"
          }
        },
        "description": "Request model for validation error demonstration"
      }
    }
  },
...
}